+++ src/TransCelerate.SDR.RuleEngine/Utilities/Helpers/ValidationRuleHelpers.cs
@@ -2,7 +2,7 @@ using FluentValidation;
 using TransCelerate.SDR.Core.Utilities.Common;
 using TransCelerate.SDR.Core.Utilities.Helpers;
 
-namespace TransCelerate.SDR.RuleEngineV5.Common
+namespace TransCelerate.SDR.RuleEngineV5.Utilities.Helpers
 {
     /// <summary>
     /// Extension helper methods for validation rules
